separate namer and hasher to another package
